Move `Resolve` function to `networking` package
