+++ internal/pkg/networking/resolve.go
@@ -1,4 +1,4 @@
-package common
+package networking
 
 import (
 	"fmt"
